restructure lfp and ripple
