+++ src/spyglass/lfp/analysis/v1/__init__.py
@@ -1,0 +1,1 @@
+from spyglass.lfp.analysis.v1 import LFPBandV1, LFPBandSelection
