+++ src/spyglass/ripple/__init__.py
@@ -1,0 +1,1 @@
+from spyglass.ripple import RippleParameters, RippleLFPSelection, RippleTimesV1
